requirements.txt
aniso8601==8.0.0
APScheduler==3.6.3
attrs==20.2.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.3
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
Pillow==8.0.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.17.3
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.20
tzlocal==2.1
Werkzeug==1.0.1
